// src/types/enums.ts
export enum EnumRoles {
  BUYER = 'BUYER',
  SELLER = 'SELLER',
  ADMIN = 'ADMIN',
}

export enum EnumLanguage {
  sh = 'sh',
  pt = 'pt',
  ru = 'ru',
  en = 'en',
  ar = 'ar',
  es = 'es',
}

export enum EnumPlatforms {
  PC = 'PC',
  PS = 'PS',
  XBOX = 'XBOX',
}

export enum EnumPayment {
  BankCard = 'Bank Card',
  CryptoCurrency = 'Crypto Currency',
  PayPal = 'PayPal',
}

export enum EnumRatingColors {
  green = '#00A190',
  lightGreen = '#00B945',
  yellow = '#FFA500',
  orange = '#FF6C00',
  red = '#DD0055',
}

export enum EnumPaymentStatus {
  SUCCEEDED = 'SUCCEEDED',
  PENDING = 'PENDING',
  FAILED = 'FAILED',
  CANCELED = 'CANCELED',
}